pyproject.toml
[project]
name = "health-tracker"
version = "0.1.0"
description = "Personal health & fitness analytics platform"
requires-python = ">=3.11"
dependencies = [
    # Backend
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20",
    "alembic>=1.14",
    "pydantic>=2.10",
    "pydantic-settings>=2.6",

    # Data source clients
    "httpx>=0.28",
    "authlib>=1.3",

    # LLM Providers
    "anthropic>=0.95",
    "openai>=1.50",
    "google-generativeai>=0.8",

    # Scheduling
    "apscheduler>=3.10,<4.0",

    # Bots
    "python-telegram-bot>=21.0",
    "discord.py>=2.3",

    # Utilities
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[tool.setuptools.packages.find]
include = ["backend*", "bot*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
